flask-backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
pymongo==3.12.0
flask-jwt-extended==4.4.4
bcrypt==3.2.0
python-dotenv==0.19.0
weaviate-client==3.26.7
langchain-google-genai==1.0.8
google-generativeai==0.7.2
langchain-community==0.2.12  